requirements.txt
Django~=4.2
Markdown~=3.4.3
Pillow
django-taggit